pyproject.toml
[tool.poetry]
name = "em"
version = "0.1.0"
description = "Startup script for emacsclient"
authors = ["J. W. Smith <jwsmith2spam at gmail dot com>"]

[tool.poetry.scripts]
em = 'em.console:run'

[tool.poetry.dependencies]
python = ">=3.6,<3.9"
PySide2 = "^5.15.0"
Pillow = "^7.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
